test(sentry): scrub_eventsテストを3分割し関連ドキュメントを更新 (#543)
* test(sentry): scrub_eventsテストを3分割し関連ドキュメントを更新

【背景】test_sentry_scrub_events.py（1731行）の肥大化により責務別分割が必要だった。

【変更内容】transaction/spanのスクラブテストをtest_sentry_transaction_span.pyへ、例外フィールド・内部タグ・extrasの補助関数テストをtest_sentry_scrub_event_helpers.pyへ抽出。events側のmodule docstringに相互参照を追加し、test_sentry_init.pyとSerenaメモリの記載を実態（6ファイル構成）に合わせて更新。

品質ゲート: 1443 tests passed / 97.60% coverage / ruff 0 errors / mypy 0 errors

* test(sentry): 非dict frameテストのモック手法を統一 (PR #543 review)

【背景】_scrub_exception_field の非dict frameテストのみ別モジュールの
private実装(sentry_primitives._logger.warning)を直接パッチしており、
他の同種テストと一貫性がなく、ロギング実装変更で無関係に壊れるリスクがあった。

【変更内容】パッチ対象をモジュールスコープの _safe_log_warning に統一し、
イベント名の assert を追加。未使用となった sentry_primitives インポートも削除。

品質ゲート: 1443 tests passed / 97.60% coverage / ruff 0 errors / mypy 0 errors

Refs #543 (PR)

---------
