+++ .serena/memories/sentry_integration.md
@@ -1,6 +1,6 @@
 # Sentry統合ガイド
 
-*最終更新: 2026年5月26日*
+*最終更新: 2026年07月27日*
 *用途: Sentry SDK設定・機密データ保護・MCP統合*
 *アクセス頻度: 低（Sentry設定・デバッグ時のみ）*
 
@@ -133,9 +133,10 @@ Sentry MCPサーバーでClaude Codeからエラー調査可能:
 ## テスト
 
 ```bash
-# Sentry統合テスト（4ファイル / 315 collected cases）
+# Sentry統合テスト（6ファイル）
 uv run pytest tests/unit/test_sentry_init.py tests/unit/test_sentry_scrub_primitives.py \
-    tests/unit/test_sentry_scrub_values.py tests/unit/test_sentry_scrub_events.py -v
+    tests/unit/test_sentry_scrub_values.py tests/unit/test_sentry_scrub_events.py \
+    tests/unit/test_sentry_transaction_span.py tests/unit/test_sentry_scrub_event_helpers.py -v
 ```
 
 ---
