+++ tests/unit/test_sentry_scrub_events.py
@@ -1,6 +1,10 @@
 """Sentryイベント整形（events）のテスト
 
 _before_send を中心としたイベント整形・例外スクラブ・再帰ガードをカバー。
+
+transaction / span のスクラブは test_sentry_transaction_span.py、
+例外フィールド・内部タグ・extras の補助関数の直接テストは
+test_sentry_scrub_event_helpers.py 側にある。
 """
 
 from __future__ import annotations
@@ -18,8 +22,6 @@ import utils.sentry_scrub_values as sentry_values
 from utils.sentry_scrub_events import (
     _SCRUBBED_EVENT_FIELDS,
     _before_send,
-    _scrub_exception_field,
-    _scrub_exception_value_item,
 )
 from utils.sentry_scrub_values import MAX_SCRUB_DEPTH
 
@@ -40,230 +42,6 @@ def test_scrubbed_event_fields_match_expected_contract() -> None:
             "exception",
         }
     )
-
-
-class TestScrubExceptionField:
-    """Sentry exception フィールドの fail-open 分岐を検証する。"""
-
-    def test_non_dict_frame_logs_warning_and_passes_through(self) -> None:
-        """非 dict frame は fail-open しつつ、無音にせず warning を残す。"""
-        exception_value = {
-            "values": [
-                {
-                    "stacktrace": {
-                        "frames": [
-                            "raw-frame",
-                            {"vars": {"password": "secret", "safe": "ok"}},
-                        ],
-                    },
-                },
-            ],
-        }
-
-        with patch.object(sentry_primitives._logger, "warning") as mock_warning:
-            result = _scrub_exception_field(exception_value)
-
-        frames = result["values"][0]["stacktrace"]["frames"]
-        assert frames[0] == "raw-frame"
-        assert frames[1]["vars"] == {"password": "[REDACTED]", "safe": "ok"}
-
-        mock_warning.assert_called_once()
-        call_kwargs = mock_warning.call_args[1]
-        assert call_kwargs["actual_type"] == "str"
-        assert call_kwargs["action"] == "skip_frame_scrub"
-
-    def test_scrubs_exception_value_string(self) -> None:
-        exception_value = {
-            "values": [
-                {
-                    "value": "DatabaseError: Connection to postgres://user:password@host/db failed",
-                },
-            ],
-        }
-
-        result = _scrub_exception_field(exception_value)
-
-        assert result["values"][0]["value"] == "[REDACTED]"
-
-    def test_scrub_exception_field_with_dict_values(self) -> None:
-        """values が dict 型の場合も _scrub_sensitive_data でkey-based scrubされる。
-
-        list-path と異なり value フィールドの明示的 [REDACTED] 置換は行わない。
-        _scrub_sensitive_data のキーベース判定に依存する（dict構造では value が
-        キー名として直接出現しないため）。
-        """
-        exception_value = {
-            "values": {
-                "error_1": {"value": "safe_data", "token": "secret_123"},
-                "error_2": {"value": "also_safe", "password": "pw_leak"},
-            },
-        }
-
-        result = _scrub_exception_field(exception_value)
-
-        # 機密キーは [REDACTED] に置換される
-        assert result["values"]["error_1"]["token"] == "[REDACTED]"  # noqa: S105
-        assert result["values"]["error_2"]["password"] == "[REDACTED]"  # noqa: S105
-        # 非機密キー・非機密値はそのまま保持
-        assert result["values"]["error_1"]["value"] == "safe_data"
-        assert result["values"]["error_2"]["value"] == "also_safe"
-        # 元の exception_value は変更されない
-        assert exception_value["values"]["error_1"]["token"] == "secret_123"  # noqa: S105
-
-    def test_scrub_exception_field_with_empty_dict_values(self) -> None:
-        exception_value: dict[str, Any] = {"values": {}}
-
-        result = _scrub_exception_field(exception_value)
-
-        assert result == {"values": {}}
-
-    def test_scrub_exception_field_without_values_key_no_warning(self) -> None:
-        """ "values" キー未存在は Sentry 仕様上の有効構造のため WARNING を出さない。
-
-        Sentry exception interface は values を必須としない
-        (getsentry/sentry interfaces/exception.py: get_path(data, "values",
-        default=[]) で空リスト扱い)。誤検知 WARNING を抑制しつつ、他キーは
-        ベストエフォートで機密スクラブを継続することを検証する。
-        """
-        exception_value: dict[str, Any] = {"type": "ValueError", "token": "secret_xyz"}  # noqa: S106
-
-        with patch("utils.sentry_scrub_events._safe_log_warning") as mock_warn:
-            result = _scrub_exception_field(exception_value)
-
-        # "values" キー未存在は正常構造のため WARNING は出力されない
-        mock_warn.assert_not_called()
-        # 機密キーはベストエフォートでスクラブされる
-        assert result["token"] == "[REDACTED]"  # noqa: S105
-        # 非機密キーは保持される
-        assert result["type"] == "ValueError"
-        # 元の exception_value は変更されない
-        assert exception_value["token"] == "secret_xyz"  # noqa: S105
-
-
-class TestHasInternalTag:
-    """_has_internal_tag ヘルパーの defensive branch coverage テスト
-
-    helper の全 branch (dict / list / fall-through) を直接 unit test 化し、
-    `_before_send` 経由の間接検証では gate 化できない defensive branch
-    (None / str / int / 空 dict / 空 list / value mismatch) の robustness を
-    回帰防止する。
-    """
-
-    def test_dict_form_with_internal_tag_returns_true(self) -> None:
-        tags = {sentry_events._INTERNAL_TAG_KEY: sentry_events._INTERNAL_TAG_VALUE}
-        assert sentry_events._has_internal_tag(tags) is True
-
-    def test_list_form_with_internal_tag_returns_true(self) -> None:
-        tags = [
-            ("env", "prod"),
-            (sentry_events._INTERNAL_TAG_KEY, sentry_events._INTERNAL_TAG_VALUE),
-        ]
-        assert sentry_events._has_internal_tag(tags) is True
-
-    @pytest.mark.parametrize(
-        "non_match_value",
-        [
-            None,
-            "string-tags",
-            42,
-            0,
-            False,
-            {},  # 空 dict
-            [],  # 空 list
-            {"other_key": "other_value"},  # dict 形式だが内部 key 不在
-            {sentry_events._INTERNAL_TAG_KEY: "wrong_value"},  # dict 形式 + value mismatch
-            [("env", "prod"), ("other_key", "other_value")],  # list 形式だが内部 tag 不在
-            [(sentry_events._INTERNAL_TAG_KEY, "wrong_value")],  # list 形式 + value mismatch
-            [("not-a-tuple-with-2-elements",)],  # list 内 malformed item
-        ],
-    )
-    def test_fall_through_returns_false(self, non_match_value: Any) -> None:
-        """dict / list 以外、または内部 tag 不在の場合は常に False を返す。
-
-        defensive fall-through branch を gate 化することで、helper の
-        unintended True 返却 (recursion guard 誤発火 → 通常 event の
-        scrub スキップ = PII 漏洩経路) を回帰防止する。
-        """
-        assert sentry_events._has_internal_tag(non_match_value) is False
-
-
-class TestScrubExceptionFailOpenBranches:
-    """_scrub_exception_frame / _scrub_exception_stacktrace の fail-open 警告分岐の直接テスト
-
-    これらの分岐 (frame 非 dict / frame_vars 非 dict・None / frames 非 list・None) は
-    従来 _before_send 経由の統合テストでしか到達せず、PII 漏洩防止に関わる重要パスのため
-    直接ユニットテストで個別に検証する。
-    """
-
-    def test_frame_not_dict_returns_input_and_warns_high_risk(self) -> None:
-        """frame が dict でない場合: 入力を破壊せず返し、HIGH リスク警告を出す (fail-open)。"""
-        with patch.object(sentry_events, "_safe_log_warning") as mock_warn:
-            result = sentry_events._scrub_exception_frame("not-a-frame")
-
-        assert result == "not-a-frame"
-        mock_warn.assert_called_once()
-        assert mock_warn.call_args.args[0] == "sentry_exception_frame_unexpected_type"
-        assert mock_warn.call_args.kwargs.get("pii_leak_risk") == "HIGH"
-
-    def test_frame_vars_unexpected_type_skips_scrub_and_warns(self) -> None:
-        """frame['vars'] が dict/None 以外: vars を素通しし警告を出す (scrub スキップ)。"""
-        frame = {"function": "handler", "vars": "not-a-dict"}
-        with patch.object(sentry_events, "_safe_log_warning") as mock_warn:
-            result = sentry_events._scrub_exception_frame(frame)
-
-        assert result["vars"] == "not-a-dict"
-        mock_warn.assert_called_once()
-        assert mock_warn.call_args.args[0] == "sentry_exception_frame_vars_unexpected_type"
-
-    def test_frame_vars_none_no_warning(self) -> None:
-        """frame['vars'] が None (正常: vars 欠如): 警告を出さず frame を返す (誤検知防止)。"""
-        frame = {"function": "handler"}
-        with patch.object(sentry_events, "_safe_log_warning") as mock_warn:
-            result = sentry_events._scrub_exception_frame(frame)
-
-        assert result == {"function": "handler"}
-        mock_warn.assert_not_called()
-
-    def test_frame_vars_dict_is_scrubbed(self) -> None:
-        frame = {"vars": {"password": "secret", "user_input": "safe_value"}}
-        result = sentry_events._scrub_exception_frame(frame)
-
-        assert result["vars"]["password"] == "[REDACTED]"  # noqa: S105
-        assert result["vars"]["user_input"] == "safe_value"
-
-    def test_frame_source_context_is_removed_and_vars_are_scrubbed(self) -> None:
-        frame = {
-            "pre_context": ["API_KEY = 'secret'"],
-            "context_line": "raise RuntimeError(password)",
-            "post_context": ["logger.info(user_email)"],
-            "vars": {"password": "secret", "user_input": "safe_value"},
-        }
-
-        result = sentry_events._scrub_exception_frame(frame)
-
-        assert "pre_context" not in result
-        assert "context_line" not in result
-        assert "post_context" not in result
-        assert result["vars"]["password"] == "[REDACTED]"  # noqa: S105
-        assert result["vars"]["user_input"] == "safe_value"
-
-    def test_frames_unexpected_type_returns_input_and_warns(self) -> None:
-        """stacktrace['frames'] が list/None 以外: stacktrace を素通しし警告を出す (fail-open)。"""
-        stacktrace = {"frames": "not-a-list"}
-        with patch.object(sentry_events, "_safe_log_warning") as mock_warn:
-            result = sentry_events._scrub_exception_stacktrace(stacktrace)
-
-        assert result == {"frames": "not-a-list"}
-        mock_warn.assert_called_once()
-        assert mock_warn.call_args.args[0] == "sentry_exception_frames_unexpected_type"
-
-    def test_frames_none_no_warning(self) -> None:
-        stacktrace = {"registers": {"rax": "0x0"}}
-        with patch.object(sentry_events, "_safe_log_warning") as mock_warn:
-            result = sentry_events._scrub_exception_stacktrace(stacktrace)
-
-        assert result == {"registers": {"rax": "0x0"}}
-        mock_warn.assert_not_called()
 
 
 class TestBeforeSend:
@@ -1249,193 +1027,6 @@ class TestBeforeSend:
         assert result["safe"] == "keep_this"  # 非機密キーは保持
 
 
-class TestBeforeSendTransaction:
-    """transaction イベント（before_send_transaction 経路）の scrub テスト
-
-    transaction は error と異なり top-level `spans` (list[dict]) を持つ
-    (Sentry transaction payload spec)。`before_send_transaction=_before_send`
-    配線により error と同一 scrub 経路を通り、span 内の機密キーが REDACT される
-    ことを検証する。
-    """
-
-    def _call_before_send(self, event: Event) -> dict[str, Any]:
-        result = _before_send(event, {})
-        assert result is not None, "_before_send が None を返しました（イベントが破棄されました）"
-        return cast(dict[str, Any], result)
-
-    def test_transaction_spans_sensitive_keys_redacted(self) -> None:
-        event = cast(
-            Event,
-            {
-                "type": "transaction",
-                "transaction": "/api/users",
-                "spans": [
-                    {
-                        "op": "http.client",
-                        "description": "GET /api/token",
-                        "data": {
-                            "auth_token": "secret-span-a",  # noqa: S106
-                            "status_code": 200,
-                        },
-                    },
-                    {
-                        "op": "db.query",
-                        "data": {"password": "secret-span-b", "rows": 5},  # noqa: S106
-                    },
-                ],
-            },
-        )
-        result_dict = self._call_before_send(event)
-        assert result_dict["spans"][0]["data"]["auth_token"] == "[REDACTED]"  # noqa: S105
-        assert result_dict["spans"][1]["data"]["password"] == "[REDACTED]"  # noqa: S105
-        # 非機密のデバッグ情報は保持される
-        assert result_dict["spans"][0]["data"]["status_code"] == 200
-        assert result_dict["spans"][1]["data"]["rows"] == 5
-        assert result_dict["spans"][0]["op"] == "http.client"
-
-    def test_transaction_span_description_query_string_redacted(self) -> None:
-        """transaction span description の query PII を REDACT する。"""
-        event = cast(
-            Event,
-            {
-                "type": "transaction",
-                "spans": [
-                    {
-                        "op": "http.client",
-                        "description": (
-                            "GET https://api.example.com/users/jane@example.com"
-                            "?token=secret123&safe=1#secret-fragment"
-                        ),
-                        "data": {},
-                    }
-                ],
-            },
-        )
-
-        result_dict = self._call_before_send(event)
-
-        description = result_dict["spans"][0]["description"]
-        assert "secret123" not in description
-        assert "jane@example.com" not in description
-        assert "secret-fragment" not in description
-        assert "token=%5BREDACTED%5D" in description
-        assert "safe=1" in description
-
-    def test_transaction_internal_tag_skips_scrub_no_recursion(self) -> None:
-        """内部通知タグ付き transaction は scrub をスキップ通過する（再帰防止）。
-
-        `tags` は base Event の top-level 属性であり transaction にも継承されるため
-        (Sentry event payload spec)、`_has_internal_tag` が transaction-shaped event
-        でも発火し、before_send_transaction 経由の無限再帰を遮断する。
-        """
-        event = cast(
-            Event,
-            {
-                "type": "transaction",
-                "tags": {
-                    sentry_events._INTERNAL_TAG_KEY: sentry_events._INTERNAL_TAG_VALUE,
-                },
-                "spans": [{"op": "db", "data": {"password": "not-scrubbed"}}],  # noqa: S106
-            },
-        )
-        result_dict = self._call_before_send(event)
-        # スキップ通過のため span は scrub されず原形のまま（再帰防止ガードの証明）
-        assert result_dict["spans"][0]["data"]["password"] == "not-scrubbed"  # noqa: S105
-
-    def test_error_event_without_spans_unaffected(self) -> None:
-        """spans を持たない error イベントは "spans" 追加の影響を受けない（回帰防止）。"""
-        event = cast(
-            Event,
-            {"request": {"headers": {"Cookie": "session=abc123"}}},
-        )
-        result_dict = self._call_before_send(event)
-        assert "spans" not in result_dict
-        assert result_dict["request"]["headers"]["Cookie"] == "[REDACTED]"
-
-    def test_transaction_request_field_scrubbed_via_transaction_path(self) -> None:
-        """transaction の top-level `request` も before_send_transaction 経路で scrub される。
-
-        WSGI/ASGI 統合では transaction イベントにも `request` (headers/query_string/url)
-        が付与される (Sentry公式)。span data より auth header 等の実 PII を含みやすいため、
-        新規配線した before_send_transaction=_before_send 経路で request 既存 scrub ロジック
-        (L745-770) が transaction-shaped event でも発火することを経験的に検証する。
-        """
-        event = cast(
-            Event,
-            {
-                "type": "transaction",
-                "transaction": "/api/users",
-                "request": {
-                    "headers": {"Authorization": "Bearer secret-xyz"},
-                    "query_string": "token=leak123&page=2",
-                    "url": "https://api.example.com/users?token=leak123",
-                },
-                "spans": [{"op": "http", "data": {"api_key": "sk-secret"}}],  # noqa: S106
-            },
-        )
-        result_dict = self._call_before_send(event)
-        # request の auth header が REDACT される
-        assert result_dict["request"]["headers"]["Authorization"] == "[REDACTED]"
-        # query_string 内のトークンが scrub される（_scrub_request_query_string 経路）
-        assert "leak123" not in result_dict["request"]["query_string"]
-        # url のクエリトークンが scrub される（_scrub_url 経路）
-        assert "leak123" not in result_dict["request"]["url"]
-        # span data も同時に scrub される
-        assert result_dict["spans"][0]["data"]["api_key"] == "[REDACTED]"  # noqa: S105
-
-    def test_transaction_structurally_valid_after_scrub(self) -> None:
-        """scrub 後も transaction が構造的に有効なまま（Relay の silent drop 防止）。
-
-        `_before_send` は `contexts` / `spans` を in-place scrub するため、
-        transaction 必須フィールド（`type` / `start_timestamp` /
-        `contexts.trace` の `trace_id` / `span_id`）が破壊されないことを保証する。
-        これらは `_is_sensitive_key` 非該当のため REDACT されず原形を保つべき。
-        破壊されると Relay が transaction を無言ドロップし、性能監視が機能しなくなる。
-        """
-        event = cast(
-            Event,
-            {
-                "type": "transaction",
-                "transaction": "/api/users",
-                "start_timestamp": 1588601261.481961,
-                "timestamp": 1588601261.488901,
-                "contexts": {
-                    "trace": {
-                        "trace_id": "1e57b752bc6e4544bbaa246cd1d05dee",
-                        "span_id": "b01b9f6349558cd1",
-                        "op": "http.server",
-                        # 機密キーは scrub されるが trace 構造は保持される
-                        "data": {"auth_token": "secret"},  # noqa: S106
-                    },
-                },
-                "spans": [
-                    {
-                        "op": "db",
-                        "span_id": "aaaa1111bbbb2222",
-                        "trace_id": "1e57b752bc6e4544bbaa246cd1d05dee",
-                        "data": {"password": "secret"},  # noqa: S106
-                    }
-                ],
-            },
-        )
-        result_dict = self._call_before_send(event)
-        # 必須トップレベルフィールドが保持される
-        assert result_dict["type"] == "transaction"
-        assert result_dict["start_timestamp"] == 1588601261.481961
-        # contexts.trace の識別子が破壊されない（非機密キーは原形保持）
-        trace = result_dict["contexts"]["trace"]
-        assert trace["trace_id"] == "1e57b752bc6e4544bbaa246cd1d05dee"
-        assert trace["span_id"] == "b01b9f6349558cd1"
-        assert trace["op"] == "http.server"
-        # trace.data 内の機密キーは scrub される
-        assert trace["data"]["auth_token"] == "[REDACTED]"  # noqa: S105
-        # spans は依然 list で識別子が保持される
-        assert isinstance(result_dict["spans"], list)
-        assert result_dict["spans"][0]["span_id"] == "aaaa1111bbbb2222"
-        assert result_dict["spans"][0]["trace_id"] == "1e57b752bc6e4544bbaa246cd1d05dee"
-        assert result_dict["spans"][0]["data"]["password"] == "[REDACTED]"  # noqa: S105
-
-
 class TestSentryProcessorBeforeSendChain:
     """logger._sentry_processor → sentry_scrub_events._before_send PII フィルター連鎖テスト
 
@@ -1534,160 +1125,6 @@ class TestSentryProcessorBeforeSendChain:
         assert user["ip_address"] == "[REDACTED]"
         assert contexts["auth"]["token"] == "[REDACTED]"  # noqa: S105
         assert contexts["auth"]["role"] == "admin"
-
-
-def test_internal_tag_value_is_valid_hex() -> None:
-    import re
-
-    from utils.sentry_scrub_events import _INTERNAL_TAG_VALUE
-
-    assert re.fullmatch(r"[0-9a-f]{32}", _INTERNAL_TAG_VALUE) is not None
-    assert len(_INTERNAL_TAG_VALUE) <= 200  # Sentry SDK tag value 上限 200文字以内
-
-
-def test_exception_value_item_redacts_dict_bytes_value() -> None:
-    result = _scrub_exception_value_item({"value": b"password=secret"})
-
-    assert result == {"value": "[REDACTED]"}
-
-
-class TestScrubSpanItem:
-    """_scrub_span_item の直接単体テスト（T-2）。
-
-    _before_send 経由の間接テストではカバーされないエッジケースを
-    直接呼び出しで検証する。
-    """
-
-    def test_non_dict_input_list_delegates_to_scrub_list_item(self) -> None:
-        """(a) 非dict入力（list）は _scrub_list_item に委譲される。
-
-        平坦な文字列リストは機密キーコンテキストを持たないためそのまま保持される。
-        """
-        result = sentry_events._scrub_span_item(["foo", "bar"], _depth=0)
-        assert result == ["foo", "bar"]
-
-    def test_non_dict_scalar_delegates_to_scrub_list_item(self) -> None:
-        result = sentry_events._scrub_span_item("plain-string", _depth=0)
-        assert result == "plain-string"
-
-    def test_max_depth_exceeded_returns_sentinel(self) -> None:
-        """(b) _depth >= MAX_SCRUB_DEPTH のとき "[MAX_DEPTH_EXCEEDED]" を返す。"""
-        result = sentry_events._scrub_span_item(
-            {"description": "GET /path"}, _depth=MAX_SCRUB_DEPTH
-        )
-        assert result == "[MAX_DEPTH_EXCEEDED]"
-
-    def test_description_without_space_and_query_scrubs_whole_via_scrub_url(self) -> None:
-        """(c) descriptionにスペースなし + '?' → 全体を _scrub_url で処理しmethod prefixなし。
-
-        "https://api.example.com/users?token=x" はスペースで分割されないため
-        value_to_scrub = description 全体。'?' を含むので _scrub_url 経路。
-        クエリの token 値が除去され、スキームとホストは保持される。
-        """
-        item = {"description": "https://api.example.com/users?token=x"}
-        result = sentry_events._scrub_span_item(item, _depth=0)
-
-        result_description = result["description"]
-        assert result_description.startswith("https://")
-        assert "token=x" not in result_description
-
-    def test_description_with_space_and_query_scrubs_via_scrub_url_preserves_method(self) -> None:
-        """(d) description が "GET /path?token=secret" 形式 → _scrub_url 経路でmethod保持。
-
-        スペースで分割後 method="GET", target="/path?token=secret"。
-        '?' を含むので _scrub_url が呼ばれ、クエリがスクラブされる。
-        "GET " プレフィックスは保持される。
-        """
-        item = {"description": "GET /path?token=secret"}
-        result = sentry_events._scrub_span_item(item, _depth=0)
-
-        result_description = result["description"]
-        assert result_description.startswith("GET ")
-        assert "secret" not in result_description
-
-    def test_description_path_email_pii_redacted_via_path_pii_pattern(self) -> None:
-        """(e) description のpath内メールPII → _PATH_PII_PATTERN で [REDACTED] 置換。
-
-        "GET /users/foo@example.com/profile" はスペースで分割後
-        target="/users/foo@example.com/profile"。'?' も '#' も含まないため
-        _PATH_PII_PATTERN.sub("[REDACTED]", ...) 経路。
-        メールアドレスが [REDACTED] に置換され、"GET " プレフィックスと残りのパスは保持。
-        """
-        item = {"description": "GET /users/foo@example.com/profile"}
-        result = sentry_events._scrub_span_item(item, _depth=0)
-
-        result_description = result["description"]
-        assert result_description.startswith("GET ")
-        assert "foo@example.com" not in result_description
-        assert "[REDACTED]" in result_description
-        assert "/profile" in result_description
-
-
-class TestSetInternalExtras:
-    """_set_internal_extras の許可リスト強制。
-
-    内部タグ付きイベントは _before_send の scrub をバイパスして Sentry に到達するため、
-    extra に書けるキーを _INTERNAL_EVENT_EXTRA_KEYS で固定し、PII を含み得るキーの
-    混入を fail-fast で拒否する多層防御 (CWE-312) を検証する。
-    """
-
-    def test_unauthorized_key_raises_value_error(self) -> None:
-        """許可リスト外キーは ValueError で拒否され、scope へ一切書き込まない。"""
-        mock_scope = MagicMock()
-
-        with pytest.raises(
-            ValueError,
-            match=r"Unauthorized key in internal event extra: 'user_email'",
-        ):
-            sentry_events._set_internal_extras(mock_scope, {"user_email": "x@example.com"})
-
-        # 拒否時は部分書き込みが起きない（原子性）
-        mock_scope.set_extra.assert_not_called()
-
-    def test_mixed_unauthorized_key_raises_before_partial_write(self) -> None:
-        """許可キー混在時も未許可キーがあれば scope へ一切書き込まない。"""
-        mock_scope = MagicMock()
-
-        with pytest.raises(
-            ValueError,
-            match=r"Unauthorized key in internal event extra: 'user_email'",
-        ):
-            sentry_events._set_internal_extras(
-                mock_scope,
-                {"error_type": "ValueError", "user_email": "x@example.com"},
-            )
-
-        mock_scope.set_extra.assert_not_called()
-
-    def test_authorized_keys_are_delegated_to_scope(self) -> None:
-        """許可リスト内キーは全て scope.set_extra へ委譲される。
-
-        実呼び出し元 _emit_scrub_failure_to_sentry は 4 キー全てを使うため、
-        いずれかの転送が欠落するリグレッションを検出できるよう全キーを検証する。
-        """
-        mock_scope = MagicMock()
-
-        sentry_events._set_internal_extras(
-            mock_scope,
-            {
-                "error_type": "ValueError",
-                "error_module": "builtins",
-                "action": "scrub_failed",
-                "event_id": "evt-1",
-            },
-        )
-
-        mock_scope.set_extra.assert_any_call("error_type", "ValueError")
-        mock_scope.set_extra.assert_any_call("error_module", "builtins")
-        mock_scope.set_extra.assert_any_call("action", "scrub_failed")
-        mock_scope.set_extra.assert_any_call("event_id", "evt-1")
-        # 許可キーのみが委譲され、余分な書き込みがない
-        assert mock_scope.set_extra.call_count == 4
-
-    def test_empty_dict_is_accepted(self) -> None:
-        mock_scope = MagicMock()
-        sentry_events._set_internal_extras(mock_scope, {})
-        mock_scope.set_extra.assert_not_called()
 
 
 def test_before_send_logger_error_receives_event_id() -> None:
